refact: move cloudsmith validation into `validations.py`
